% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_connection.R
\name{duckdb_connection}
\alias{duckdb_connection}
\alias{duckdb_connection,ParquetArray-method}
\alias{duckdb_connection,ParquetArraySeed-method}
\alias{duckdb_connection,ParquetColumn-method}
\alias{duckdb_connection,ParquetFactTable-method}
\alias{duckdb_connection,ParquetMatrix-method}
\title{DuckDB Connection Accessor}
\usage{
duckdb_connection(x, ...)
}
\arguments{
\item{x}{An object to get the DuckDB connection value.}

\item{...}{Additional arguments, for use in specific methods.}
}
\description{
Get the DuckDB connection value contained in an object.
}
\examples{
duckdb_connection
showMethods("duckdb_connection")

}
\author{
Patrick Aboyoun
}
