% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBArray-utils.R
\name{DuckDBArray-utils}
\alias{DuckDBArray-utils}
\alias{Ops,DuckDBArray,DuckDBArray-method}
\alias{Ops,DuckDBArray,atomic-method}
\alias{Ops,atomic,DuckDBArray-method}
\alias{Math,DuckDBArray-method}
\alias{is.finite,DuckDBArray-method}
\alias{is.infinite,DuckDBArray-method}
\alias{is.nan,DuckDBArray-method}
\alias{rowSums,DuckDBArray-method}
\alias{colSums,DuckDBArray-method}
\alias{is_nonzero,DuckDBArray-method}
\alias{nzcount,DuckDBArray-method}
\title{Common operations on DuckDBArray objects}
\description{
Common operations on \linkS4class{DuckDBArray} objects.
}
\section{Group Generics}{

DuckDBTable objects have support for S4 group generic functionality:
\describe{
  \item{\code{Arith}}{\code{"+"}, \code{"-"}, \code{"*"}, \code{"^"},
    \code{"\%\%"}, \code{"\%/\%"}, \code{"/"}}
  \item{\code{Compare}}{\code{"=="}, \code{">"}, \code{"<"}, \code{"!="},
    \code{"<="}, \code{">="}}
  \item{\code{Logic}}{\code{"&"}, \code{"|"}}
  \item{\code{Ops}}{\code{"Arith"}, \code{"Compare"}, \code{"Logic"}}
  \item{\code{Math}}{\code{"abs"}, \code{"sign"}, \code{"sqrt"},
    \code{"ceiling"}, \code{"floor"}, \code{"trunc"}, \code{"log"},
    \code{"log10"}, \code{"log2"}, \code{"acos"}, \code{"acosh"},
    \code{"asin"}, \code{"asinh"}, \code{"atan"}, \code{"atanh"},
    \code{"exp"}, \code{"expm1"}, \code{"cos"}, \code{"cosh"},
    \code{"sin"}, \code{"sinh"}, \code{"tan"}, \code{"tanh"},
    \code{"gamma"}, \code{"lgamma"}}
  \item{\code{Summary}}{\code{"max"}, \code{"min"}, \code{"range"},
    \code{"prod"}, \code{"sum"}, \code{"any"}, \code{"all"}}
 }
 See \link[methods]{S4groupGeneric} for more details.
}

\section{Numerical Data Methods}{

In the code snippets below, \code{x} is a DuckDBArray object:
\describe{
  \item{\code{is.finite(x)}:}{
    Returns a DuckDBArray containing logicals that indicate which values are
    finite.
  }
  \item{\code{is.infinite(x)}:}{
    Returns a DuckDBArray containing logicals that indicate which values are
    infinite.
  }
  \item{\code{is.nan(x)}:}{
    Returns a DuckDBArray containing logicals that indicate which values are
    Not a Number.
  }
  \item{\code{rowSums(x, dims = 1)}:}{
    Calculates the row sums of \code{x}.
    \describe{
      \item{\code{dims}}{An integer specifying which dimensions to sum over,
        namely \code{dims + 1}, \ldots.}
    }
  }
  \item{\code{colSums(x, dims = 1)}:}{
    Calculates the column sums of \code{x}.
    \describe{
      \item{\code{dims}}{An integer specifying which dimensions to sum over,
        namely \code{1:dims}.}
    }
  }
}
}

\section{Sparisty Methods}{

In the code snippets below, \code{x} is a DuckDBArray object:
\describe{
  \item{\code{is_nonzero(x)}:}{
    Returns a DuckDBArray containing logicals that indicate if the values in
    each of the columns of \code{x} are non-zero.
  }
  \item{\code{nzcount(x)}:}{
    Returns the total number of non-zero values.
  }
  \item{\code{is_sparse(x)}:}{
    Defined as \code{(ncol(x) == 1L) && ((nzcount(x) / nrow(x)) < 0.5)}.
  }
}
}

\author{
Patrick Aboyoun
}
\keyword{methods}
\keyword{utilities}
