% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBArray.R
\name{DuckDBArray}
\alias{DuckDBArray}
\alias{DuckDBArray-class}
\alias{[,DuckDBArray,ANY,ANY,ANY-method}
\alias{aperm,DuckDBArray-method}
\alias{dbconn,DuckDBArray-method}
\alias{is_nonzero,DuckDBArray-method}
\alias{nzcount,DuckDBArray-method}
\alias{t,DuckDBArray-method}
\alias{type,DuckDBArray-method}
\alias{type<-,DuckDBArray-method}
\alias{Ops,DuckDBArray,DuckDBArray-method}
\alias{Ops,DuckDBArray,atomic-method}
\alias{Ops,atomic,DuckDBArray-method}
\alias{Math,DuckDBArray-method}
\alias{Summary,DuckDBArray-method}
\alias{mean,DuckDBArray-method}
\alias{median.DuckDBArray}
\alias{quantile.DuckDBArray}
\alias{var,DuckDBArray,ANY-method}
\alias{sd,DuckDBArray-method}
\alias{mad,DuckDBArray-method}
\title{DuckDB tables as DelayedArray objects}
\usage{
DuckDBArray(conn, keycols, datacols, type = NULL, ...)
}
\arguments{
\item{conn}{Either a string containing the path to the data files or a
\code{tbl_duckdb_connection} object.}

\item{keycols}{Either a character vector or a list of character vectors
containing the names of the columns that comprise the primary key.}

\item{datacols}{String containing the name of the column in the DuckDB table
that specifies the value of the array.}

\item{type}{String specifying the type of the data values; one of
\code{"logical"}, \code{"integer"}, \code{"integer64"}, \code{"double"}, or
\code{"character"}. If \code{NULL}, it is determined by inspecting the data.}

\item{...}{Further arguments to be passed to \code{read_parquet}.}
}
\description{
The DuckDBArray class is a \link[DelayedArray]{DelayedArray} subclass
for representing and operating on a DuckDB table.

All the operations available for \link[DelayedArray]{DelayedArray}
objects work on DuckDBArray objects.
}
\examples{
# Create a data.frame from the Titanic data
df <- do.call(expand.grid, c(dimnames(Titanic), stringsAsFactors = FALSE))
df$fate <- Titanic[as.matrix(df)]

# Write data to a parquet file
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(df, tf)

pqarray <- DuckDBArray(tf, keycols = c("Class", "Sex", "Age", "Survived"), datacols = "fate")

}
\seealso{
\code{\link{DuckDBArraySeed}},
\code{\link[DelayedArray]{DelayedArray}}
}
\author{
Patrick Aboyoun
}
