% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBColumn-class.R
\name{DuckDBColumn-class}
\alias{DuckDBColumn-class}
\alias{dbconn,DuckDBColumn-method}
\alias{tblconn,DuckDBColumn-method}
\alias{length,DuckDBColumn-method}
\alias{names,DuckDBColumn-method}
\alias{type,DuckDBColumn-method}
\alias{type<-,DuckDBColumn-method}
\alias{extractROWS,DuckDBColumn,ANY-method}
\alias{head,DuckDBColumn-method}
\alias{tail,DuckDBColumn-method}
\alias{as.vector,DuckDBColumn-method}
\alias{show,DuckDBColumn-method}
\alias{showAsCell,DuckDBColumn-method}
\title{DuckDBColumn objects}
\description{
DuckDBColumn objects
}
\author{
Patrick Aboyoun
}
