% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBColumn-utils.R
\name{DuckDBColumn-utils}
\alias{DuckDBColumn-utils}
\alias{Ops,DuckDBColumn,DuckDBColumn-method}
\alias{Ops,DuckDBColumn,atomic-method}
\alias{Ops,atomic,DuckDBColumn-method}
\alias{Math,DuckDBColumn-method}
\alias{Summary,DuckDBColumn-method}
\alias{is.finite,DuckDBColumn-method}
\alias{is.infinite,DuckDBColumn-method}
\alias{is.nan,DuckDBColumn-method}
\alias{mean,DuckDBColumn-method}
\alias{var,DuckDBColumn,ANY-method}
\alias{sd,DuckDBColumn-method}
\alias{median.DuckDBColumn}
\alias{quantile.DuckDBColumn}
\alias{mad,DuckDBColumn-method}
\alias{IQR,DuckDBColumn-method}
\alias{unique,DuckDBColumn-method}
\alias{\%in\%,DuckDBColumn,ANY-method}
\alias{table,DuckDBColumn-method}
\alias{is_nonzero,DuckDBColumn-method}
\alias{nzcount,DuckDBColumn-method}
\title{DuckDBColumn Utilities}
\description{
DuckDBColumn Utilities
}
\author{
Patrick Aboyoun
}
