% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBColumn.R
\name{DuckDBColumn}
\alias{DuckDBColumn}
\alias{DuckDBColumn-class}
\alias{\%in\%,DuckDBColumn,ANY-method}
\alias{as.vector,DuckDBColumn-method}
\alias{dbconn,DuckDBColumn-method}
\alias{extractROWS,DuckDBColumn,ANY-method}
\alias{head,DuckDBColumn-method}
\alias{is_nonzero,DuckDBColumn-method}
\alias{is.finite,DuckDBColumn-method}
\alias{is.infinite,DuckDBColumn-method}
\alias{is.nan,DuckDBColumn-method}
\alias{length,DuckDBColumn-method}
\alias{names,DuckDBColumn-method}
\alias{nzcount,DuckDBColumn-method}
\alias{show,DuckDBColumn-method}
\alias{showAsCell,DuckDBColumn-method}
\alias{tail,DuckDBColumn-method}
\alias{tblconn,DuckDBColumn-method}
\alias{type,DuckDBColumn-method}
\alias{type<-,DuckDBColumn-method}
\alias{unique,DuckDBColumn-method}
\alias{Ops,DuckDBColumn,DuckDBColumn-method}
\alias{Ops,DuckDBColumn,atomic-method}
\alias{Ops,atomic,DuckDBColumn-method}
\alias{Math,DuckDBColumn-method}
\alias{Summary,DuckDBColumn-method}
\alias{mean,DuckDBColumn-method}
\alias{median.DuckDBColumn}
\alias{quantile.DuckDBColumn}
\alias{var,DuckDBColumn,ANY-method}
\alias{sd,DuckDBColumn-method}
\alias{mad,DuckDBColumn-method}
\alias{IQR,DuckDBColumn-method}
\alias{table,DuckDBColumn-method}
\title{DuckDBColumn objects}
\description{
DuckDBColumn objects
}
\author{
Patrick Aboyoun
}
