% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBConnection.R
\name{DuckDBConnection}
\alias{DuckDBConnection}
\alias{acquireDuckDBConn}
\alias{releaseDuckDBConn}
\title{Acquire and Release the DuckDB Connection}
\usage{
acquireDuckDBConn(conn = dbConnect(duckdb(bigint = "integer64")))

releaseDuckDBConn()
}
\arguments{
\item{conn}{A \code{duckdb_connection} object.}
}
\value{
\code{acquireDuckDBConn} returns the current DuckDB connection.
}
\description{
Acquire and Release the DuckDB connection used by the
\pkg{BiocDuckDB} package.
}
\details{
\code{acquireDuckDBConn} will add a \code{duckdb_connection} object to
the \pkg{BiocDuckDB} package cache.
\code{releaseDuckDBConn} will remove the \code{duckdb_connection} object
from that cache.
}
\examples{
\dontrun{
acquireDuckDBConn()
releaseDuckDBConn()
}

}
\author{
Patrick Aboyoun
}
