% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBDataFrameList.R
\name{DuckDBDataFrameList}
\alias{DuckDBDataFrameList}
\alias{DuckDBDataFrameList-class}
\alias{split,DuckDBDataFrame,DuckDBColumn-method}
\title{DuckDB-backed DataFrameList}
\description{
Create a DuckDB-backed \linkS4class{DataFrameList} object.
}
\author{
Patrick Aboyoun
}
