% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBList-class.R
\name{DuckDBList-class}
\alias{DuckDBList-class}
\alias{length,DuckDBList-method}
\alias{names,DuckDBList-method}
\alias{names<-,DuckDBList-method}
\alias{elementNROWS,DuckDBList-method}
\alias{unlist,DuckDBList-method}
\alias{extractROWS,DuckDBList,ANY-method}
\alias{getListElement,DuckDBList-method}
\alias{head,DuckDBList-method}
\alias{tail,DuckDBList-method}
\title{DuckDB-backed List}
\description{
Create a DuckDB-backed \linkS4class{List} object.
}
\author{
Patrick Aboyoun
}
