% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBMatrix.R
\name{DuckDBMatrix}
\alias{DuckDBMatrix}
\alias{DuckDBMatrix-class}
\alias{[,DuckDBMatrix,ANY,ANY,ANY-method}
\alias{matrixClass,DuckDBArray-method}
\title{DuckDB tables as DelayedMatrix objects}
\usage{
DuckDBMatrix(con, row, col, datacols, keycols = c(row, col), type = NULL, ...)
}
\arguments{
\item{con}{Either a string containing the path to the underlying data files
or a \code{tbl_duckdb_connection} object.}

\item{row}{Either a character vector or a named list of character vectors
containing the names of the columns in the DuckDB table that specify the
rows of the matrix.}

\item{col}{Either a character vector or a named list of character vectors
containing the names of the columns in the DuckDB table that specify the
columns of the matrix.}

\item{datacols}{String containing the name of the column in the DuckDB table
that specifies the value of the matrix.}

\item{keycols}{Either a character vector or a list of character vectors
containing the names of the columns in the DuckDB table that specify the
rows and columns of the matrix.}

\item{type}{String specifying the type of the data values; one of
\code{"logical"}, \code{"integer"}, \code{"integer64"}, \code{"double"}, or
\code{"character"}. If \code{NULL}, it is determined by inspecting the data.}

\item{...}{Further arguments to be passed to \code{read_parquet}.}
}
\description{
The DuckDBMatrix class is a \link[DelayedArray]{DelayedMatrix} subclass
for representing and operating on a DuckDB table.

All the operations available for \link[DelayedArray]{DelayedMatrix}
objects work on DuckDBMatrix objects.
}
\examples{
# Create a data.frame from a matrix
df <- data.frame(
  rowname = rep(rownames(state.x77), times = ncol(state.x77)),
  colname = rep(colnames(state.x77), each = nrow(state.x77)),
  value = as.vector(state.x77)
)

# Write data to a parquet file
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(df, tf)

pqmat <- DuckDBMatrix(tf, row = "rowname", col = "colname", datacols = "value")

}
\author{
Patrick Aboyoun
}
