% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBTable-utils.R
\name{DuckDBTable-utils}
\alias{DuckDBTable-utils}
\alias{Ops,DuckDBTable,DuckDBTable-method}
\alias{Ops,DuckDBTable,atomic-method}
\alias{Ops,atomic,DuckDBTable-method}
\alias{Math,DuckDBTable-method}
\alias{Summary,DuckDBTable-method}
\alias{is.finite,DuckDBTable-method}
\alias{is.infinite,DuckDBTable-method}
\alias{is.nan,DuckDBTable-method}
\alias{mean,DuckDBTable-method}
\alias{var,DuckDBTable,ANY-method}
\alias{sd,DuckDBTable-method}
\alias{median.DuckDBTable}
\alias{quantile.DuckDBTable}
\alias{mad,DuckDBTable-method}
\alias{IQR,DuckDBTable-method}
\alias{rowSums,DuckDBTable-method}
\alias{colSums,DuckDBTable-method}
\alias{unique,DuckDBTable-method}
\alias{\%in\%,DuckDBTable,ANY-method}
\alias{table,DuckDBTable-method}
\alias{is_nonzero,DuckDBTable-method}
\alias{nzcount,DuckDBTable-method}
\alias{is_sparse,DuckDBTable-method}
\title{DuckDBTable Utilities}
\description{
DuckDBTable Utilities
}
\author{
Patrick Aboyoun
}
