% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keynames.R
\name{keynames}
\alias{keynames}
\alias{keydimnames}
\alias{keydimnames<-}
\alias{nkey}
\alias{nkeydim}
\alias{keynames,DuckDBTable-method}
\alias{keydimnames,DuckDBTable-method}
\alias{keydimnames<-,DuckDBTable-method}
\alias{nkey,DuckDBTable-method}
\alias{nkeydim,DuckDBTable-method}
\title{Key Names and Key Count}
\usage{
keynames(x)

keydimnames(x, value)

keydimnames(x) <- value

nkey(x)

nkeydim(x)
}
\arguments{
\item{x}{An object to get the keycols related information.}

\item{value}{A character vector of keycols dimension names.}
}
\description{
Get the keycols names, keycols dimension names, or keycols count of an object.
}
\examples{
keynames
showMethods("keynames")

keydimnames
showMethods("keydimnames")

nkey
showMethods("nkey")

}
\author{
Patrick Aboyoun
}
