% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LanguageList.R
\name{LanguageList}
\alias{LanguageList}
\alias{LanguageList-class}
\title{Lists of Language Objects}
\usage{
LanguageList(...)
}
\arguments{
\item{\dots}{language objects, possibly named.}
}
\description{
An extension of \code{\linkS4class{SimpleList}} that holds \code{language} objects.
}
\examples{
getClassDef("language")
LanguageList(jenny = quote(867-5309))
LanguageList(x = quote(x), sqrtx = quote(sqrt(x)), logx = quote(log(x)))

}
