% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetArray.R
\name{ParquetArray}
\alias{ParquetArray}
\alias{ParquetArray-class}
\alias{[,ParquetArray,ANY,ANY,ANY-method}
\alias{aperm,ParquetArray-method}
\alias{dbconn,ParquetArray-method}
\alias{is_nonzero,ParquetArray-method}
\alias{nzcount,ParquetArray-method}
\alias{t,ParquetArray-method}
\alias{type,ParquetArray-method}
\alias{type<-,ParquetArray-method}
\alias{Ops,ParquetArray,ParquetArray-method}
\alias{Ops,ParquetArray,atomic-method}
\alias{Ops,atomic,ParquetArray-method}
\alias{Math,ParquetArray-method}
\alias{Summary,ParquetArray-method}
\alias{mean,ParquetArray-method}
\alias{median.ParquetArray}
\alias{quantile.ParquetArray}
\alias{var,ParquetArray,ANY-method}
\alias{sd,ParquetArray-method}
\alias{mad,ParquetArray-method}
\title{Parquet datasets as DelayedArray objects}
\usage{
ParquetArray(conn, key, fact, type = NULL, ...)
}
\arguments{
\item{conn}{Either a string containing the path to the data files or a
\code{tbl_duckdb_connection} object.}

\item{key}{Either a character vector or a list of character vectors
containing the names of the columns that comprise the primary key.}

\item{fact}{String containing the name of the column in the Parquet data
that specifies the value of the array.}

\item{type}{String specifying the type of the Parquet data values;
one of \code{"logical"}, \code{"integer"}, \code{"double"}, or
\code{"character"}. If \code{NULL}, this is determined by inspecting
the data.}

\item{...}{Further arguments to be passed to \code{read_parquet}.}
}
\description{
The ParquetArray class is a \link[DelayedArray]{DelayedArray} subclass
for representing and operating on a Parquet dataset.

All the operations available for \link[DelayedArray]{DelayedArray}
objects work on ParquetArray objects.
}
\examples{
# Create a data.frame from the Titanic data
df <- do.call(expand.grid, c(dimnames(Titanic), stringsAsFactors = FALSE))
df$fate <- Titanic[as.matrix(df)]

# Write data to a parquet file
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(df, tf)

pqarray <- ParquetArray(tf, key = c("Class", "Sex", "Age", "Survived"), fact = "fate")

}
\seealso{
\code{\link{DuckDBArraySeed}},
\code{\link[DelayedArray]{DelayedArray}}
}
\author{
Patrick Aboyoun
}
