% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParquetFactTable.R
\name{ParquetFactTable}
\alias{ParquetFactTable}
\alias{ParquetFactTable-class}
\alias{[,ParquetFactTable,ANY,ANY,ANY-method}
\alias{all.equal.ParquetFactTable}
\alias{as.data.frame,ParquetFactTable-method}
\alias{bindCOLS,ParquetFactTable-method}
\alias{colnames,ParquetFactTable-method}
\alias{colnames<-,ParquetFactTable-method}
\alias{coltypes}
\alias{coltypes,ParquetFactTable-method}
\alias{coltypes<-}
\alias{coltypes<-,ParquetFactTable-method}
\alias{dbconn,ParquetFactTable-method}
\alias{ncol,ParquetFactTable-method}
\alias{nrow,ParquetFactTable-method}
\alias{rownames,ParquetFactTable-method}
\alias{Ops,ParquetFactTable,ParquetFactTable-method}
\alias{Ops,ParquetFactTable,atomic-method}
\alias{Ops,atomic,ParquetFactTable-method}
\alias{Math,ParquetFactTable-method}
\alias{Summary,ParquetFactTable-method}
\alias{mean,ParquetFactTable-method}
\alias{median.ParquetFactTable}
\alias{quantile.ParquetFactTable}
\alias{var,ParquetFactTable,ANY-method}
\alias{sd,ParquetFactTable-method}
\alias{mad,ParquetFactTable-method}
\title{ParquetFactTable objects}
\usage{
ParquetFactTable(
  conn,
  key,
  fact = setdiff(colnames(conn), names(key)),
  type = NULL,
  ...
)
}
\arguments{
\item{conn}{Either a string containing the path to the Parquet data or a
\code{tbl_duckdb_connection} object.}

\item{key}{Either a character vector or a named list of character vectors
containing the names of the columns in the Parquet data that specify
the primary key of the array.}

\item{fact}{Either a character vector containing the names of the columns
in the Parquet data that specify the facts}

\item{type}{An optional named character vector where the names specify the
column names and the values specify the column type; one of
\code{"logical"}, \code{"integer"}, \code{"double"}, or \code{"character"}.}

\item{...}{Further arguments to be passed to \code{read_parquet}.}
}
\description{
ParquetFactTable is a low-level helper class for representing a
pointer to a Parquet fact table.
}
\examples{
# Create a data.frame from the Titanic data
df <- do.call(expand.grid, c(dimnames(Titanic), stringsAsFactors = FALSE))
df$fate <- as.integer(Titanic[as.matrix(df)])

# Write data to a parquet file
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(df, tf)

tbl <- ParquetFactTable(tf, key = c("Class", "Sex", "Age", "Survived"), fact = "fate")

}
\author{
Patrick Aboyoun
}
