% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tblconn.R
\name{tblconn}
\alias{tblconn}
\title{Accessing DB Table Information}
\usage{
tblconn(x, filter = TRUE)
}
\arguments{
\item{x}{An object with a database table connection.}

\item{filter}{A logical value indicating whether to apply key column
filtering to the database table connection.}
}
\value{
returns a database table connection object.
}
\description{
Get a database table connection object
}
\examples{
tblconn
showMethods("tableconn")

}
\author{
Patrick Aboyoun
}
