% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBArray-utils.R
\name{DuckDBArray-utils}
\alias{DuckDBArray-utils}
\alias{Ops,DuckDBArray,DuckDBArray-method}
\alias{Ops,DuckDBArray,atomic-method}
\alias{Ops,atomic,DuckDBArray-method}
\alias{Math,DuckDBArray-method}
\alias{is.finite,DuckDBArray-method}
\alias{is.infinite,DuckDBArray-method}
\alias{is.nan,DuckDBArray-method}
\alias{rowSums,DuckDBArray-method}
\alias{colSums,DuckDBArray-method}
\alias{\%in\%,DuckDBArray,ANY-method}
\alias{is_nonzero,DuckDBArray-method}
\alias{nzcount,DuckDBArray-method}
\title{DuckDBArray Utilities}
\description{
DuckDBArray Utilities
}
\author{
Patrick Aboyoun
}
