% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuckDBArraySeed.R
\name{DuckDBArraySeed}
\alias{DuckDBArraySeed}
\alias{DuckDBArraySeed-class}
\alias{[,DuckDBArraySeed,ANY,ANY,ANY-method}
\alias{aperm,DuckDBArraySeed-method}
\alias{dbconn,DuckDBArraySeed-method}
\alias{DelayedArray,DuckDBArraySeed-method}
\alias{dim,DuckDBArraySeed-method}
\alias{dimnames,DuckDBArraySeed-method}
\alias{extract_array,DuckDBArraySeed-method}
\alias{extract_sparse_array,DuckDBArraySeed-method}
\alias{is_nonzero,DuckDBArraySeed-method}
\alias{is_sparse,DuckDBArraySeed-method}
\alias{nzcount,DuckDBArraySeed-method}
\alias{t,DuckDBArraySeed-method}
\alias{type,DuckDBArraySeed-method}
\alias{type<-,DuckDBArraySeed-method}
\alias{Ops,DuckDBArraySeed,DuckDBArraySeed-method}
\alias{Ops,DuckDBArraySeed,atomic-method}
\alias{Ops,atomic,DuckDBArraySeed-method}
\alias{Math,DuckDBArraySeed-method}
\alias{Summary,DuckDBArraySeed-method}
\alias{mean,DuckDBArraySeed-method}
\alias{median.DuckDBArraySeed}
\alias{quantile.DuckDBArraySeed}
\alias{var,DuckDBArraySeed,ANY-method}
\alias{sd,DuckDBArraySeed-method}
\alias{mad,DuckDBArraySeed-method}
\title{DuckDBArraySeed objects}
\usage{
DuckDBArraySeed(conn, key, fact, type = NULL, ...)
}
\arguments{
\item{conn}{Either a string containing the path to the data files or a
\code{tbl_duckdb_connection} object.}

\item{key}{Either a character vector or a list of character vectors
containing the names of the columns that comprise the primary key.}

\item{fact}{String containing the name of the column in the Parquet data
that specifies the value of the array.}

\item{type}{String specifying the type of the Parquet data values;
one of \code{"logical"}, \code{"integer"}, \code{"double"}, or
\code{"character"}. If \code{NULL}, this is determined by inspecting
the data.}

\item{...}{Further arguments to be passed to \code{read_parquet}.}
}
\description{
DuckDBArraySeed is a low-level helper class for representing a
pointer to a DuckDB table.

Note that a DuckDBArraySeed object is not intended to be used directly.
Most end users will typically create and manipulate a higher-level
\link{ParquetArray} object instead. See \code{?\link{ParquetArray}} for
more information.
}
\examples{
# Create a data.frame from the Titanic data
df <- do.call(expand.grid, c(dimnames(Titanic), stringsAsFactors = FALSE))
df$fate <- as.integer(Titanic[as.matrix(df)])

# Write data to a parquet file
tf <- tempfile()
on.exit(unlink(tf))
arrow::write_parquet(df, tf)

pqaseed <- DuckDBArraySeed(tf, key = c("Class", "Sex", "Age", "Survived"), fact = "fate")

}
\seealso{
\code{\link{ParquetArray}},
\code{\link[S4Arrays]{Array}}
}
\author{
Patrick Aboyoun
}
