% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeArray.R
\name{writeArray}
\alias{writeArray}
\title{Write an Array-Like Object}
\usage{
writeArray(
  x,
  path,
  dim_names = names(dimnames(x)) \%||\% sprintf("dim\%d", seq_along(dim(x))),
  value_name = "value",
  partition_size = dim(x),
  partition_names = sprintf("\%s_group", dim_names),
  ...
)
}
\arguments{
\item{x}{An array-like object.}

\item{path}{The path to write the array-like object to.}

\item{dim_names}{A character vector of names for the dimensions of the array.}

\item{value_name}{The name for the column containing the array values.}

\item{partition_size}{A vector of integers specifying the sub-array size of
each partition. The default is to write the entire array as a single
partition.}

\item{partition_names}{A character vector of names for the partitions of the
array.}

\item{...}{Additional arguments to pass to \code{arrow::write_dataset}.}
}
\description{
An \code{arrow::write_dataset} wrapper function to write array-like objects.
}
\examples{
# Write the Titanic dataset to a single Parquet file
tf1 <- tempfile()
writeArray(Titanic, tf1)
list.files(tf1, full.names = TRUE, recursive = TRUE)

# Write the Titanic dataset to multiple Parquet files
tf2 <- tempfile()
writeArray(Titanic, tf2, partition_size = c(2, 2, 2, 2))
list.files(tf2, full.names = TRUE, recursive = TRUE)

# Write the Titanic dataset to multiple csv files
tf3 <- tempfile()
writeArray(Titanic, tf3, partition_size = c(2, 2, 2, 2), format = "csv")
list.files(tf3, full.names = TRUE, recursive = TRUE)

# Write a random array to multiple Parquet files
tf4 <- tempfile()
X <- randomSparseMatrix(50000, 2^17, density = 0.0001)
writeArray(X, tf4, partition_size = c(nrow(X), 100000))
files <- list.files(tf4, full.names = TRUE, recursive = TRUE)
files <- setNames(sprintf("\%.1f MB", file.size(files) / 2^20), files)
files

}
\author{
Patrick Aboyoun
}
